Separate API wrappers and initializers into different files
