+++ src/utils/initializers.ts
@@ -1,0 +1,12 @@
+export const initRun = () => {
+    return {
+        gameSlug: "soulsilver",
+        prevLocationSlug: "new-bark-town",
+        starterSlug: "",
+        encounterList: [],
+        caughtPokemonSlugsList: [],
+        numDead: 0,
+        numCheckpoints: 10,
+        numCheckpointsCleared: 0,
+    };
+};
