+++ tsconfig.json
@@ -1,6 +1,5 @@
 {
     "compilerOptions": {
-        "target": "es5",
         "lib": ["dom", "dom.iterable", "esnext"],
         "allowJs": true,
         "skipLibCheck": true,
